[Issue-620] Multiple use of the nested FlinkSerializer
